introduce topics component
